requirements.txt
flask
gunicorn
openai
pypdfium2>=4.30
Pillow
requests
python-dotenv